+++ cmd/lhdiff/main.go
@@ -2,8 +2,10 @@ package main
 
 import (
 	"flag"
+	"fmt"
 	"github.com/SmartBear/lhdiff"
 	"io/ioutil"
+	"os"
 )
 
 func main() {
@@ -13,7 +15,15 @@ func main() {
 	rightFile := flag.Arg(1)
 	left, _ := ioutil.ReadFile(leftFile)
 	right, _ := ioutil.ReadFile(rightFile)
-	mappings := lhdiff.Lhdiff(string(left), string(right), 4, !*compact)
-	lhdiff.PrintMappings(mappings)
-}
+	mappings, err := lhdiff.Lhdiff(string(left), string(right), 4, !*compact)
 
+	if err != nil {
+		_, _ = fmt.Fprintln(os.Stderr, err.Error())
+		os.Exit(1)
+	}
+	err = lhdiff.PrintMappings(mappings)
+	if err != nil {
+		_, _ = fmt.Fprintln(os.Stderr, err.Error())
+		os.Exit(1)
+	}
+}